requirements.txt
fastapi==0.115.6
langchain_huggingface==0.1.2
langchain_postgres==0.0.12
PyPDF2==3.0.1
psycopg2==2.9.10
python-multipart==0.0.20
load_dotenv